update colour(s) to use go-roster
